MM-46374 : Move "Themes" type of MM-redux to another location
