+++ components/plugin_marketplace/marketplace_list/navigation_row/navigation_row.test.tsx
@@ -4,7 +4,7 @@
 import React from 'react';
 import {shallow} from 'enzyme';
 
-import {Theme} from 'mattermost-redux/types/themes';
+import {Theme} from 'mattermost-redux/selectors/entities/preferences';
 
 import NavigationRow, {NavigationRowProps} from './navigation_row';
