+++ components/post_view/message_attachments/action_button/action_button.tsx
@@ -4,7 +4,7 @@
 import React from 'react';
 
 import {changeOpacity} from 'mattermost-redux/utils/theme_utils';
-import {Theme} from 'mattermost-redux/types/themes';
+import {Theme} from 'mattermost-redux/selectors/entities/preferences';
 import {PostAction, PostActionOption} from '@mattermost/types/integration_actions';
 
 import LoadingWrapper from 'components/widgets/loading/loading_wrapper';
