+++ packages/mattermost-redux/src/utils/theme_utils.ts
@@ -1,8 +1,8 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import {Theme, ThemeTypeMap} from 'mattermost-redux/types/themes';
-import {Preferences} from '../constants';
+import {LegacyThemeType, Theme, ThemeKey, ThemeType} from 'mattermost-redux/selectors/entities/preferences';
+import {Preferences} from 'mattermost-redux/constants';
 
 export function makeStyleFromTheme(getStyleFromTheme: (a: any) => any): (a: any) => any {
     let lastTheme: any;
@@ -115,6 +115,8 @@ export const blendColors = (background: string, foreground: string, opacity: num
 
     return `rgba(${red},${green},${blue},${alpha})`;
 };
+
+type ThemeTypeMap = Record<ThemeType | LegacyThemeType, ThemeKey>;
 
 // object mapping theme types to their respective keys for retrieving the source themes directly
 // - supports mapping old themes to new themes
