+++ plugins/channel_header_plug/channel_header_plug.test.tsx
@@ -4,7 +4,7 @@
 import React from 'react';
 
 import {Channel, ChannelMembership} from '@mattermost/types/channels';
-import {Theme} from 'mattermost-redux/types/themes';
+import {Theme} from 'mattermost-redux/selectors/entities/preferences';
 
 import ChannelHeaderPlug from 'plugins/channel_header_plug/channel_header_plug';
 import {mountWithIntl} from '../../tests/helpers/intl-test-helper';
