+++ plugins/mobile_channel_header_plug/mobile_channel_header_plug.test.tsx
@@ -10,7 +10,7 @@ import {mountWithIntl} from 'tests/helpers/intl-test-helper';
 import {createCallContext} from 'utils/apps';
 import {AppCallResponseTypes} from 'mattermost-redux/constants/apps';
 import {Channel, ChannelMembership} from '@mattermost/types/channels';
-import {Theme} from 'mattermost-redux/types/themes';
+import {Theme} from 'mattermost-redux/selectors/entities/preferences';
 
 describe('plugins/MobileChannelHeaderPlug', () => {
     const testPlug = {
