+++ plugins/mobile_channel_header_plug/mobile_channel_header_plug.tsx
@@ -8,7 +8,7 @@ import {injectIntl, IntlShape} from 'react-intl';
 import {AppCallResponseTypes} from 'mattermost-redux/constants/apps';
 import {AppBinding} from '@mattermost/types/apps';
 import {Channel, ChannelMembership} from '@mattermost/types/channels';
-import {Theme} from 'mattermost-redux/types/themes';
+import {Theme} from 'mattermost-redux/selectors/entities/preferences';
 
 import {PluginComponent} from 'types/store/plugins';
 import {createCallContext} from 'utils/apps';
